Improve support for 32-bit builds on Windows
This change allows mapping large files with offsets exceeding 32-bit boundary

+++ include/mio/mmap.hpp
@@ -57,7 +57,7 @@ template<access_mode AccessMode, typename ByteT>
 struct basic_mmap
 {
     using value_type = ByteT;
-    using size_type = size_t;
+    using size_type = int64_t;
     using reference = value_type&;
     using const_reference = const value_type&;
     using pointer = value_type*;
